Move integration models into imbi.automations
This clears up an occasional cyclic import problem that I found when
writing some tests.
